fix(deserialization): Use a specific deserializer
In the Endpoint definition, "healthcheck" is an object.
But to keep legacy compatibility, we must handle the fact that sometimes, "healthcheck" is a boolean.
The try with Jackson annotation was not successful, so we go back to a specific deserializer.

This should have no impact on the swagger.json in the Management API.

Fixes gravitee-io/issues#5570